Swagger-UI

+++ src/main/java/com/matchvagas/backend/dto/CidadeRequestDTO.java
@@ -1,6 +1,3 @@
 package com.matchvagas.backend.dto;
 
-public record CidadeRequestDTO(
-    String nome,
-    Integer estadoId
-) {}
+public record CidadeRequestDTO(String nome, Long estadoId) {}
