+++ src/main/java/com/matchvagas/backend/dto/EstadoResponseDTO.java
@@ -3,10 +3,4 @@ package com.matchvagas.backend.dto;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public record EstadoResponseDTO(
-    Integer id,
-    String nome,
-    String uf,
-    Integer paisId,
-    String paisNome
-) {}
+public record EstadoResponseDTO(Long id, String nome, String uf, Long paisId, String paisNome) {}
